+++ src/commands/build-web-app.js
@@ -1,0 +1,22 @@
+import path from 'path'
+
+import webpackBuild from '../webpackBuild'
+
+export default function(args, cb) {
+  require('./clean-app')(args)
+
+  console.log(`nwb: build-web-app`)
+  webpackBuild(args, {
+    devtool: 'source-map',
+    entry: {
+      app: path.resolve('src/index.js')
+    },
+    output: {
+      filename: '[name].js',
+      path: path.resolve('public/build')
+    },
+    plugins: {
+      vendorChunkName: 'vendor'
+    }
+  }, cb)
+}