+++ src/constants.js
@@ -2,6 +2,7 @@ export const REACT_VERSION = '~0.14.0'
 
 export const REACT_APP = 'react-app'
 export const REACT_COMPONENT = 'react-component'
+export const WEB_APP = 'web-app'
 export const WEB_MODULE = 'web-module'
 
-export const PROJECT_TYPES = [REACT_APP, REACT_COMPONENT, WEB_MODULE]
+export const PROJECT_TYPES = [REACT_APP, REACT_COMPONENT, WEB_APP, WEB_MODULE]