+++ templates/web-app/nwb.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  // Let nwb know this is a web app when generic build commands are used
+  type: 'web-app'
+}