+++ templates/web-app/src/index.js
@@ -1,0 +1,3 @@
+let app = document.querySelector('#app')
+
+app.innerHTML = '<h2>Welcome to {{name}}</h2>'